% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_label_to_plot.R
\name{add_label_to_plot}
\alias{add_label_to_plot}
\title{Add information about critical values to a plot.}
\usage{
add_label_to_plot(.plot, x, y, add_tick = FALSE, label, parse = TRUE,
  vjust = 2, size = 5, colour = "black")
}
\arguments{
\item{.plot}{The plot upon which the information should be added.}

\item{x}{A number that gives the first coordinate of the point
where information in \code{.label} should be added.}

\item{y}{A number (default value \code{0}) that gives the second
axis coordinate of the point where the information in
\code{.label} should be added.}

\item{add_tick}{Logical value, default value \code{FALSE}, that can
be used to add a "tick" to the plot at the coordinate given by
\code{x} and \code{y}.}

\item{label}{The text to be used as a label, based at the position
given by \code{x} and \code{y}.}

\item{parse}{Logical value, default value \code{TRUE}, which
decides if the text given in \code{label} should be parsed.}

\item{vjust}{A number (default value \code{2}) that governs the
vertical adjustment of \code{label} relative the value
\code{y}. Note that this argument is relative to the
\code{size}-value, and an increase in \code{size} might thus
require a decrease in \code{vjust} in order to get the desired
position.}

\item{size}{The value that governs the size of the text, default
value \code{5}.}

\item{colour}{The colour, default value \code{"black"}, that will
be used for the added details.}
}
\value{
An updated version of \code{.plot}, where information has
    been added.
}
\description{
This function collects some code that helps place information about
critical values to a density plot.
}
\keyword{internal}
