% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_defaults.R
\docType{data}
\name{distribution_details}
\alias{distribution_details}
\title{Distribution defaults}
\format{An object of class \code{list} of length 5.}
\usage{
distribution_details
}
\description{
A list that describes the default values to be used for the
distributions.  Functions with a \code{distribution}-argument will
have this tested against the details given in this list.  Note that
the \code{type}-part of the \code{distribution}-argument must be
one of the top-level names of the list \code{distribution_details},
and that the parameters must be within the acceptable range.
}
\keyword{datasets}
