% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVL_defaults.R
\docType{data}
\name{HVL_defaults}
\alias{HVL_defaults}
\title{Internal default values.}
\format{An object of class \code{list} of length 6.}
\usage{
HVL_defaults
}
\description{
This is a list with some internal default-values, that will be
used as references for the different functions in the package.
}
\keyword{internal}
