% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVL_sample.R
\name{HVL_sample}
\alias{HVL_sample}
\title{Simulate samples to be used in \code{HVL_shiny},}
\usage{
HVL_sample(dist, par, n = 100, M = 100, seed = NULL)
}
\arguments{
\item{dist}{A character vector that must be one of the implemented
distributions.}

\item{par}{A list with the arguments to be used when sampling from
the distribution given in \code{dist}.}

\item{n}{An integer, default value \code{100}, that specifies the
number of observations in each individual sample.}

\item{M}{An integer, default value \code{100}, that specifies the
number of replicates of samples of size \code{n} from the
distribution specified in \code{distribution}.}

\item{seed}{An integer that can be used to specify a seed-value to
be used when sampling.  The default value \code{NULL} turns of
the part of the code that sets the seed.}
}
\value{
The \code{M} samples of size \code{n} will be returned as
    an array.
}
\description{
This function will be called from \code{HVL_shiny}, and its purpose
is to create the samples needed for some of the interactive
investigations.
}
