% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVL_server_Master.R
\name{HVL_server_Master}
\alias{HVL_server_Master}
\title{Master for the \code{server}-part of the \code{HVL_shiny}-interface.}
\usage{
HVL_server_Master(.env, .env2)
}
\arguments{
\item{.env}{The environment where the original argument to
\code{HVL_shiny} lives, i.e. the argument
\code{initial_values}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
This function is the master function that takes of the
    dynamic part of the interface for the
    \code{HVL_shiny}-function.  (The resulting updates are stored
    in the environments \code{.env} and \code{.env2}.)
}
\description{
This function takes care of the
}
\keyword{internal}
