% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVL_shiny.R
\name{HVL_shiny}
\alias{HVL_shiny}
\title{Shiny app for HVLstatistikk}
\usage{
HVL_shiny(initial_values = list(selected_tab = "info", par_list = NULL))
}
\arguments{
\item{initial_values}{A list that can be used to specify the
default values when the application starts.}
}
\value{
An interactive shiny application will be loaded.
}
\description{
An interactive interface to some simple statistical concepts.
}
