% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVL_shiny_language.R
\name{HVL_shiny_language}
\alias{HVL_shiny_language}
\title{The language-interface for \code{HVL_shiny}-interface.}
\usage{
HVL_shiny_language(.env, .env2)
}
\arguments{
\item{.env}{The environment where the original argument to
\code{HVL_shiny} lives, i.e. the argument
\code{initial_values}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
This function updates the language of the
    \code{HVL_shiny}-interface.
}
\description{
This function takes care of the
}
\keyword{internal}
