% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVL_shiny_quit.R
\name{HVL_shiny_quit}
\alias{HVL_shiny_quit}
\title{Create the button to manually quit \code{HVL_shiny}.}
\usage{
HVL_shiny_quit(.env, .env2)
}
\arguments{
\item{.env}{The environment where the original argument to
\code{HVL_shiny} lives, i.e. the argument
\code{initial_values}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
This function decides if the insist-quit button should be
    present or not, and it also takes care of the language to be
    used on this button.
}
\description{
Create the button to manually quit \code{HVL_shiny}.
}
\keyword{internal}
