% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVLstatistikk-package.r
\docType{package}
\name{HVLstatistikk}
\alias{HVLstatistikk}
\alias{HVLstatistikk-package}
\title{HVLstatistikk.}
\description{
A package with a few simple shiny-applications, intended to be used
when teaching statistics at the Western Norway Universiy of Applied
Sciences.
}
