% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ind.R
\name{Ind}
\alias{Ind}
\title{An indicator function to reveal if a given observation belong to a
specified event.}
\usage{
Ind(x, .type = "c", event, border_criteria = c("<=", "<="),
  .res_type = "L")
}
\arguments{
\item{x}{A vector of observations from some distribution.}

\item{.type}{Either \code{"c"} for continuous (default) or
\code{"d"} for discrete.  If the value is \code{"c"}, then
\code{event} must be a numeric vector of length 2, otherwise it
can be a vector of any length.}

\item{event}{A numeric vector.  The vector must have length 2 when
\code{.type="c"}, since the values then will be used as the
lower and upper limits of an interval.  The argument
\code{border_criteria} endpoints of the intervals are included
in the event.}

\item{border_criteria}{A vector of length two, that must be one the
following four: \code{c("<=", "<=")} (default), \code{c("<",
"<=")}, \code{c("<", "<")}, \code{c("<", "<=")}.  This is
needed when \code{.type="c"}, since it will govern whether or
not the endpoints of the interval should be included in event.
This argument is ignored when \code{.type="d"}.}

\item{.res_type}{Either \code{"L"} for logical (default) or
\code{"I"} for integer.  Governs the format of the resulting
vector.}
}
\value{
This function will return vector of the same length as
    \code{x}, which depending on \code{.res_type} either will be of
    type logical or of type integer.  The values will be used to
    indicate whether or not the observations in \code{x} was
    contained in the set defined by \code{event}.
}
\description{
An indicator function to reveal if a given observation belong to a
specified event.
}
\keyword{internal}
