% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLN_plot.R
\name{LLN_plot}
\alias{LLN_plot}
\title{A plot to visualise the law of large numbers}
\usage{
LLN_plot(.plot = NULL, .true_value = NULL, .indicator_values,
  .m_of_n = NULL, .colours = c("red", "black"), .ltys = c(1, 2),
  .show_text = TRUE, ...)
}
\arguments{
\item{.plot}{The plot upon which additional information should be
added.  A new plot will be created if the default value
\code{NULL} is used.}

\item{.true_value}{The true value of the event of interest.  The
default value \code{NULL} can be used to create a plot also in
the case this value is unknown.}

\item{.indicator_values}{A vector with the result of an
indicator-function that tells us whether the desired event was
observed in the different cases of interest. This could e.g. be
related to the result of a coin-toss being head.}

\item{.m_of_n}{An integer that shows how many of the
indicator-values that should be used. This is included in order
to enable an animation of the development. The default value
\code{NULL} will set this argument to be the length of the
vector given in \code{.indicator_values}.}

\item{.colours}{A character-vector, default value \code{c("red",
"black")}, that gives the \code{colour}-arguments to be used
for the lines showing \code{.indicator_values} and
\code{.true_value}.  The second part will be ignored when
\code{.true_value=NULL}.}

\item{.ltys}{An integer-valued vector, default value \code{c(1,2)},
that gives the \code{lty}-arguments to be used for the lines
showing \code{.indicator_values} and \code{.true_value}.  The
second part will be ignored when \code{.true_value=NULL}.}

\item{.show_text}{Logical value, default \code{TRUE} that will add
the true value given in \code{.true_value} (when known) and the
estimated value based on \code{.m_of_n} and
\code{indicator_values}.}

\item{...}{Additional arguments to be used in the
\code{xy_sprintf}-function.}
}
\value{
A plot to visualise the law of large numbers.
}
\description{
A plot to visualise the law of large numbers
}
