% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_plot.R
\name{prob_plot}
\alias{prob_plot}
\title{Visualisation of a probability density function.}
\usage{
prob_plot(x, y, .plot = NULL, .colour = "black", .lty = 1,
  .size = 0.7, .alpha = 0.7, .lower = NULL, .upper = NULL, ...)
}
\arguments{
\item{x}{A vector with the values to be used along the first axis.}

\item{y}{A vector with the values along the second axis.}

\item{.plot}{The plot upon which additional information should be
added.  A new plot will be created if the default value
\code{NULL} is used.}

\item{.colour}{The \code{colour}-value to be used for the graph,
default value \code{"black"}.}

\item{.lty}{The \code{lty}-value to be used for the graph, default
value \code{1}.}

\item{.size}{The \code{size}-value to be used for the graph,
default value \code{0.7}.}

\item{.alpha}{The \code{alpha}-value to be used for the graph,
default value \code{0.7}.}

\item{.lower}{Default value \code{NULL}. To be used when a lower
end of the area is desired.  If this argument is \code{NULL}
and an upper end is specified in \code{.upper}, then this
argument will be reset to the lowest value given in \code{x}.}

\item{.upper}{Default value \code{NULL}. To be used when an upper
end of the area is desired.  If this argument is \code{NULL}
and a lower end is specified in \code{.lower}, then this
argument will be reset to the highest value given in \code{x}.}

\item{...}{Other arguments passed on to the plot of the area, like
\code{fill = "red"}, \code{colour = "red"}, \code{alpha = 0.6}
or \code{size = 0.1}.}
}
\value{
A plot showing a graph (given by \code{x} and \code{y}).
    If at least one of \code{.lower} and \code{.upper} are
    different from \code{NULL}, then a shaded area representing a
    probability will be added too.  Additional graphs/areas can be
    added iteratively by the help of the \code{.plot}-argument.
}
\description{
A helper function to add a graph and mark an area corresponding to
some probability.
}
