% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_info.R
\name{quantile_info}
\alias{quantile_info}
\title{Create a list with quantile-information.}
\usage{
quantile_info(.type = c("chisq", "norm", "t"), .alpha = 0.05,
  .df = NULL, .df_txt = "nu", digits = 2, not_IEC60559 = TRUE,
  .plotmath = TRUE, .decimal = ".")
}
\arguments{
\item{.type}{The distribution to be used, which should be one of
\code{c("chisq", "norm", "t")}.}

\item{.alpha}{The number to be used for the alpha value when
computing the quantile, which as usual refers to the upper tail
of the distribution.  The default value for this argument is
\code{0.05}.}

\item{.df}{The degrees of freedom, default \code{NULL}, to be used
for those quantiles that requires this argument.}

\item{.df_txt}{The text to be used for the specification of the
df-value.  The default value is \code{"nu"}, but it might in
some cases be of interest to use e.g. \code{"n-1"} or
\code{"n-2"}.}

\item{digits}{An integer, default value \code{2}, which will decide
the number of digits to use when rounding the quantile values.
Note: The rounding will}

\item{.plotmath}{A logical value, default \code{TRUE}, which will
ensure that the format of the returned character-string works
as desired when included as a part of some text or label in a
\code{ggplot}.  In particular, this ensures that any trailing
zeros are properly preserved when \code{parse=TRUE} is used
when adding the text.}

\item{.decimal}{A character, default value \code{.}, which is to be
used for the decimal sign for the converted number.  This
enables the decimal sign to be replaced with \code{,}, which is
of interest for those countries where that is the norm.}

\item{not_IEC60599}{A logical value, default \code{TRUE}, which can
be used to round \code{x} in the same manner a human normally
would do, e.g. that "2.5" rounds to "3" instead of "2".
Reminder: This is included to avoid that students must be
explained why the rounding used in the text is different from
what they themselves have computed.}
}
\description{
The purpose of this function is to create a list with information
related to different quantiles, in order for this to be used as
arguments in \code{add_label_to_plot}.
}
\keyword{internal}
