% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_shiny.R
\name{test_shiny}
\alias{test_shiny}
\title{Shiny app to be used for some simple investigations.}
\usage{
test_shiny(distribution = list(type = "normal", mean = -2:2, sd = c(0.1,
  0.25, 0.5, 1, 2, 4, 8)), n_max = 250, R_max = 1000, seed = NULL)
}
\arguments{
\item{distribution}{A list that gives details related to the
distribution that will be used for the example of interest.
The list must contain the name of the desired distribution and
a specification of the required parameters. Note that it is
possible to give a vector of values for the parameters.}

\item{n_max}{An integer, default value \code{250} that specifies
the highest number of observations that will be sampled from
the distribution specified in \code{distribution}.}

\item{R_max}{An integer, default value \code{1000}, that specifies
the highest number of replicates of samples of size at most
\code{n_max} from the distribution specified in
\code{distribution}.}

\item{seed}{An integer that can be used to specify a seed-value to
be used when sampling (in order to obtain reproducibility).
The default value \code{NULL} turns of the part of the code
that sets the seed.}
}
\value{
An interactive shiny application will be loaded (if
    \code{data_dir} points to a relevant directory containing the
    desired data).
}
\description{
This function opens an interactive Shiny-application that can be
used to explain some of the learning outcomes.
}
