% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_plot.R
\name{the_plot}
\alias{the_plot}
\title{The plot function}
\usage{
the_plot(.input)
}
\arguments{
\item{.input}{The main argument, which is the \code{input}-list from
the internal body of the \code{shiny}-application.}
}
\value{
A plot in accordance with the nodes detected in
    \code{.input}.
}
\description{
This function creates the plots to be used in the interactive investigation.
}
