% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_formatC.R
\name{xy_formatC}
\alias{xy_formatC}
\title{A wrapper for \code{formatC}}
\usage{
xy_formatC(x, digits = 3, not_IEC60559 = TRUE, .plotmath = TRUE,
  .decimal = ".")
}
\arguments{
\item{x}{The number that should be converted to text.}

\item{digits}{A number, default value \code{3}, to be given to the
\code{digits}-argument of \code{round}.}

\item{.plotmath}{A logical value, default \code{TRUE}, which will
ensure that the format of the returned character-string works
as desired when included as a part of some text or label in a
\code{ggplot}.  In particular, this ensures that any trailing
zeros are properly preserved when \code{parse=TRUE} is used
when adding the text.}

\item{.decimal}{A character, default value \code{.}, which is to be
used for the decimal sign for the converted number.  This
enables the decimal sign to be replaced with \code{,}, which is
of interest for those countries where that is the norm.}

\item{not_IEC60599}{A logical value, default \code{TRUE}, which can
be used to round \code{x} in the same manner a human normally
would do, e.g. that "2.5" rounds to "3" instead of "2".
Reminder: This is included to avoid that students must be
explained why the rounding used in the text is different from
what they themselves have computed.}
}
\value{
A character-string based on \code{x}, modified according to
    the description given in the arguments \code{digits},
    \code{not_IEC60599}, \code{.plotmath} and \code{.decimal}.
    Reminder: The result will have an attribute named \code{equal},
    which is included sine the intended use of this function is
    that it will be called from \code{xy_sprintf}, which then will
    either add an equality sign or an approximation sign based on
    the value of the \code{equal}-attribute.
}
\description{
This wrapper for \code{formatC} includes code that allows for some
additional tweaking to be taken care of when a number is converted
to a text-string.  It was primarily created in order to resolve an
issue related to the disappearance of trailing zeros when adding
numbers to a \code{ggplot}. (Reminder: The problem occurred when
\code{parse=TRUE} was used when adding text/labels.)
}
\keyword{internal}
