% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_sprintf.R
\name{xy_sprintf}
\alias{xy_sprintf}
\title{Slightly adjusted version of sprintf}
\usage{
xy_sprintf(.txt, x, ...)
}
\arguments{
\item{.txt}{A textstring to be included at the left side of the
equality (or approximation) sign.}

\item{x}{A number which will be included at the right side of the
equality (or approximation) sign.  Note that \code{x} will be
converted to a text-string by the help of the function
\code{xy_formatC}.}

\item{...}{Arguments to be given to \code{xy_formatC}, which will
govern how \code{x} is transformed into the text to be included
at the right side of the equality (or approximation) sign.}
}
\value{
This function will return a character-string that either
    states that \code{.txt} is equal to \code{x}, or that
    \code{.txt} is approximately equal to \code{x}.  Depending on
    the arguments in \code{...}, this text can then either be used
    to present the result in a \code{ggplot} or it can be included
    in the describing text.
}
\description{
Slightly adjusted version of sprintf
}
\keyword{internal}
